UPdated tests

+++ tests/testthat/test-addPareArticle.R
@@ -1,11 +1,13 @@
 test_that("minimal", {
-  repo <- makeRepo()
-  testthat::skip_if(!R6::is.R6(repo))
+  if (checkSuggests()) {
+    repo <- makeRepo()
+    testthat::skip_if(!R6::is.R6(repo))
 
-  suppressWarnings(addPareArticle(repo))
+    suppressWarnings(addPareArticle(repo))
 
-  path <- file.path(repo$getPath(), "vignettes", "articles", "PareReport.Rmd")
+    path <- file.path(repo$getPath(), "vignettes", "articles", "PareReport.Rmd")
 
-  expect_true(file.exists(path))
-  unlink(repo$getPath(), recursive = TRUE)
+    expect_true(file.exists(path))
+    unlink(repo$getPath(), recursive = TRUE)
+  }
 })
